Add files via upload
The "Stock Prediction of Google using LSTM in Python" is an exciting and ambitious project that aims to leverage the power of Long Short-Term Memory (LSTM) networks to predict the future stock prices of Google Inc. This GitHub repository houses the complete implementation of the project, showcasing a step-by-step guide to building an efficient and accurate stock prediction model using Python and deep learning techniques.